--- src/api/notice/notice-kind.ts
@@ -1,6 +1,0 @@
-export enum NoticeKind {
-  RECRUIT = 'recruit',
-  EVENT = 'event',
-  NORMAL = 'general',
-  ACADEMIC = 'academic',
-}
