+++ src/api/notice/notice.ts
@@ -9,6 +9,13 @@ export interface NoticePaginationParams {
   limit?: number;
 }
 
+export enum NoticeKind {
+  RECRUIT = 'recruit',
+  EVENT = 'event',
+  NORMAL = 'general',
+  ACADEMIC = 'academic',
+}
+
 export interface NoticeSearchParams {
   search?: string;
   tags?: string[];
@@ -52,3 +59,33 @@ export const getNotice = async (id: number) =>
     ...data,
     deadline: data.deadline ? data.deadline : undefined,
   }));
+
+export const GET_NOTICES = gql(`
+  query GetNotices($offset: Int, $limit: Int) {
+    notices(offset: $offset, limit: $limit) {
+      list {
+        id
+        title
+        views
+        body
+        deadline
+        createdAt
+        author
+        imageUrl
+        tags {
+          id
+          name
+        }
+      }
+      total
+    }
+  }
+`);
+
+export const CREATE_NOTICE = gql(`
+  mutation CreateNotice($title: String!, $body: String!, $deadline: Date, $tags: [Int!], $images: [String!]) {
+    createNotice(title: $title, body: $body, deadline: $deadline, tags: $tags, images: $images) {
+      id
+    }
+  }
+`);
